Staging

+++ Models/AppModels/PaySlip.cs
@@ -8,7 +8,7 @@ namespace TimesheetBE.Models.AppModels
         public EmployeeInformation EmployeeInformation { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int TotalHours { get; set; }
+        public double TotalHours { get; set; }
         public double TotalAmount { get; set; }
         public string Rate { get; set; }
         public DateTime PaymentDate { get; set; }
